requirements.txt
pytest==8.0.0
pyyaml==6.0.1
pandas>=2.0.0
openpyxl>=3.1.0
watchdog>=3.0.0
